move current shields under advancedView
