+++ components/brave_extension/extension/brave_extension/components/braveShields.tsx
@@ -7,18 +7,11 @@ import * as React from 'react'
 // CSS normalizer
 import 'emptykit.css'
 
-// Feature-specific components
-import { ShieldsPanel } from 'brave-ui/features/shields'
-
 // Components group
-import Header from './header'
-import InterfaceControls from './interfaceControls'
-import PrivacyControls from './privacyControls'
-import Footer from './footer'
+import AdvancedView from './advancedView'
 
 // Types
 import { Tab, PersistentData } from '../types/state/shieldsPannelState'
-import { isShieldsEnabled, getFavicon } from '../helpers/shieldsUtils'
 import {
   ShieldsToggled,
   BlockAdsTrackers,
@@ -53,109 +46,19 @@ interface Props {
   persistentData: PersistentData
 }
 
-interface State {
-  isBlockedListOpen: boolean
-}
-
-export default class Shields extends React.PureComponent<Props, State> {
-  constructor (props: Props) {
-    super(props)
-    this.state = { isBlockedListOpen: false }
-  }
-
-  get favicon (): string {
-    const { url } = this.props.shieldsPanelTabData
-    return getFavicon(url)
-  }
-
-  get isShieldsEnabled (): boolean {
-    const { braveShields } = this.props.shieldsPanelTabData
-    return isShieldsEnabled(braveShields)
-  }
-
-  setBlockedListOpen = () => {
-    this.setState({ isBlockedListOpen: !this.state.isBlockedListOpen })
-  }
-
+export default class Shields extends React.PureComponent<Props, {}> {
   render () {
     const { shieldsPanelTabData, persistentData, actions } = this.props
-    const { isBlockedListOpen } = this.state
-
     if (!shieldsPanelTabData) {
       return null
     }
 
     return (
-      <ShieldsPanel data-test-id='brave-shields-panel' style={{ width: '370px' }}>
-        <button onClick={actions.setAdvancedViewFirstAccess.bind(this, false)}>
-          hello state is: {persistentData.isFirstAccess.toString()}
-        </button>
-        <Header
-          enabled={this.isShieldsEnabled}
-          favicon={this.favicon}
-          origin={origin}
-          hostname={shieldsPanelTabData.hostname}
-          isBlockedListOpen={isBlockedListOpen}
-          adsBlocked={shieldsPanelTabData.adsBlocked}
-          trackersBlocked={shieldsPanelTabData.trackersBlocked}
-          httpsUpgrades={shieldsPanelTabData.httpsRedirected}
-          scriptsBlocked={shieldsPanelTabData.javascriptBlocked}
-          fingerprintingBlocked={shieldsPanelTabData.fingerprintingBlocked}
-          shieldsToggled={actions.shieldsToggled}
-        />
-        {
-          this.isShieldsEnabled && (
-            <>
-              <InterfaceControls
-                // Global props
-                isBlockedListOpen={isBlockedListOpen}
-                setBlockedListOpen={this.setBlockedListOpen}
-                hostname={shieldsPanelTabData.hostname}
-                favicon={this.favicon}
-                // Ads/Trackers
-                ads={shieldsPanelTabData.ads}
-                adsBlocked={shieldsPanelTabData.adsBlocked}
-                adsBlockedResources={shieldsPanelTabData.adsBlockedResources}
-                trackers={shieldsPanelTabData.trackers}
-                trackersBlocked={shieldsPanelTabData.trackersBlocked}
-                trackersBlockedResources={shieldsPanelTabData.trackersBlockedResources}
-                blockAdsTrackers={actions.blockAdsTrackers}
-                // HTTPS Upgrades
-                httpsRedirected={shieldsPanelTabData.httpsRedirected}
-                httpUpgradableResources={shieldsPanelTabData.httpUpgradableResources}
-                httpsRedirectedResources={shieldsPanelTabData.httpsRedirectedResources}
-                httpsEverywhereToggled={actions.httpsEverywhereToggled}
-              />
-              <PrivacyControls
-                // Global props
-                isBlockedListOpen={isBlockedListOpen}
-                setBlockedListOpen={this.setBlockedListOpen}
-                hostname={shieldsPanelTabData.hostname}
-                favicon={this.favicon}
-                // JavaScript
-                javascript={shieldsPanelTabData.javascript}
-                javascriptBlocked={shieldsPanelTabData.javascriptBlocked}
-                noScriptInfo={shieldsPanelTabData.noScriptInfo}
-                blockJavaScript={actions.blockJavaScript}
-                allowScriptOriginsOnce={actions.allowScriptOriginsOnce}
-                setScriptBlockedCurrentState={actions.setScriptBlockedCurrentState}
-                setGroupedScriptsBlockedCurrentState={actions.setGroupedScriptsBlockedCurrentState}
-                setAllScriptsBlockedCurrentState={actions.setAllScriptsBlockedCurrentState}
-                setFinalScriptsBlockedState={actions.setFinalScriptsBlockedState}
-                // Cookies
-                blockCookies={actions.blockCookies}
-                cookies={shieldsPanelTabData.cookies}
-                // Fingerprinting
-                fingerprinting={shieldsPanelTabData.fingerprinting}
-                fingerprintingBlocked={shieldsPanelTabData.fingerprintingBlocked}
-                fingerprintingBlockedResources={shieldsPanelTabData.fingerprintingBlockedResources}
-                blockFingerprinting={actions.blockFingerprinting}
-              />
-            </>
-          )
-        }
-        <Footer isBlockedListOpen={isBlockedListOpen} />
-      </ShieldsPanel>
+      <AdvancedView
+        shieldsPanelTabData={shieldsPanelTabData}
+        persistentData={persistentData}
+        actions={actions}
+      />
     )
   }
 }
