+++ components/test/brave_extension/components/advancedView/controls/cookiesControl_test.tsx
@@ -3,9 +3,9 @@
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 import * as React from 'react'
-import CookiesControl, { Props } from '../../../../brave_extension/extension/brave_extension/components/controls/cookiesControl'
-import { BlockCookiesOptions } from '../../../../brave_extension/extension/brave_extension/types/other/blockTypes'
-import * as actionTypes from '../../../../brave_extension/extension/brave_extension/constants/shieldsPanelTypes'
+import CookiesControl, { Props } from '../../../../../brave_extension/extension/brave_extension/components/advancedView/controls/cookiesControl'
+import { BlockCookiesOptions } from '../../../../../brave_extension/extension/brave_extension/types/other/blockTypes'
+import * as actionTypes from '../../../../../brave_extension/extension/brave_extension/constants/shieldsPanelTypes'
 import { shallow } from 'enzyme'
 
 const fakeProps: Props = {
@@ -14,7 +14,7 @@ const fakeProps: Props = {
   blockCookies: (setting: BlockCookiesOptions) => ({ type: actionTypes.BLOCK_COOKIES, setting })
 }
 
-describe('CookiesControl component', () => {
+describe('AdvancedView CookiesControl component', () => {
   const baseComponent = (props: Props) =>
     <CookiesControl {...props} />
